Merry_mid/requirements.txt
opencv-python
mediapipe
numpy
pygame
Pillow 